Edit note button functionality, theme adjustments

+++ app/src/main/java/com/boostedpenguin/clipboardmanager/EditNoteActivity.kt
@@ -4,15 +4,25 @@ import android.content.ClipData
 import android.content.ClipboardManager
 import android.content.Context
 import android.content.Intent
+import android.graphics.BlendMode
+import android.graphics.BlendModeColorFilter
+import android.graphics.Color
 import android.os.Bundle
+import android.text.Spannable
+import android.text.SpannableString
+import android.text.style.ForegroundColorSpan
 import android.view.Menu
 import android.view.MenuItem
 import android.widget.EditText
 import android.widget.Toast
 import androidx.activity.viewModels
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
+import androidx.lifecycle.Observer
 import com.boostedpenguin.clipboardmanager.databinding.ActivityCreateNoteBinding
 import com.boostedpenguin.clipboardmanager.room.Note
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 
 class EditNoteActivity : AppCompatActivity() {
     private lateinit var binding: ActivityCreateNoteBinding
@@ -20,25 +30,28 @@ class EditNoteActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        model.currentNote = intent.getSerializableExtra("NOTE_SELECTED") as Note
+        model.currentNote.value = intent.getSerializableExtra("NOTE_SELECTED") as Note
 
         binding = ActivityCreateNoteBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
 
-        val m = model.currentNote?.content
+        val m = model.currentNote.value?.content
 
         binding.noteContent.setText(m)
         setSupportActionBar(findViewById(R.id.createToolbar))
         supportActionBar?.title = ""
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
         supportActionBar?.setDisplayShowHomeEnabled(true)
+
+        model.currentNote.observe (this, Observer {
+            invalidateOptionsMenu()
+        })
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         menuInflater.inflate(R.menu.item_view_menu, menu)
-        val item = menu.findItem(R.id.action_save)
-        item.actionView.setOnClickListener {
+        menu.findItem(R.id.action_save).actionView.setOnClickListener {
             val content = findViewById<EditText>(R.id.note_content).text.toString()
             if(!content.isNullOrEmpty()) {
                 model.update(content)
@@ -47,6 +60,13 @@ class EditNoteActivity : AppCompatActivity() {
             else {
                 Toast.makeText(this, "Can't save an empty clipboard!", Toast.LENGTH_SHORT).show()
             }
+        }
+
+        if(model.currentNote.value!!.favorite) {
+            menu.findItem(R.id.action_favorite).icon.colorFilter = BlendModeColorFilter(ContextCompat.getColor(applicationContext, R.color.buttonSave), BlendMode.SRC_IN)
+        }
+        else {
+            menu.findItem(R.id.action_favorite).icon.clearColorFilter()
         }
         return true
     }
@@ -58,12 +78,32 @@ class EditNoteActivity : AppCompatActivity() {
                 true
             }
             R.id.action_delete -> {
-                startActivity(Intent(this, MainActivity::class.java))
-                model.deleteCurrent()
+
+                val alertDialog: AlertDialog? = this.let {
+                    val builder = MaterialAlertDialogBuilder(this, R.style.AlertDialogTheme)
+                    builder.apply {
+                        setPositiveButton("DELETE"
+                        ) { _, _ ->
+                            model.deleteCurrent()
+                            startActivity(Intent(context, MainActivity::class.java))
+                        }
+                        setNegativeButton("Cancel"
+                        ) { dialog, _ ->
+                            // User cancelled the dialog
+                            dialog.cancel()
+                        }
+                        setTitle("Delete selected items")
+                    }
+                    builder.create()
+                }
+
+                alertDialog?.show()
+
                 true
             }
             R.id.action_favorite -> {
                 model.updateFavorite()
+                invalidateOptionsMenu()
                 true
             }
             R.id.action_share -> true
@@ -74,7 +114,7 @@ class EditNoteActivity : AppCompatActivity() {
     private fun copy() {
         val clipboard: ClipboardManager =
                 getSystemService(Context.CLIPBOARD_SERVICE) as ClipboardManager
-        val clip = ClipData.newPlainText("clipboard_content", model.currentNote.content)
+        val clip = ClipData.newPlainText("clipboard_content", model.currentNote.value?.content)
         clipboard.setPrimaryClip(clip)
 
         Toast.makeText(applicationContext, "Copied", Toast.LENGTH_SHORT).show()
